backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.40
llama-index>=0.9.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
openai>=1.0.0
pyyaml>=6.0
tiktoken>=0.5.0